+++ qupath-core-processing/src/main/java/qupath/opencv/ops/ImageOp.java
@@ -21,12 +21,18 @@
 
 package qupath.opencv.ops;
 
+import java.io.IOException;
+import java.net.URI;
+import java.util.Collection;
+import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 
 import org.bytedeco.opencv.opencv_core.Mat;
 
 import qupath.lib.images.servers.ImageChannel;
 import qupath.lib.images.servers.PixelType;
+import qupath.lib.io.UriResource;
 import qupath.lib.regions.Padding;
 
 /**
@@ -37,7 +43,7 @@ import qupath.lib.regions.Padding;
  * <p>
  * Operations may be chained
  */
-public interface ImageOp {
+public interface ImageOp extends UriResource {
 	
 	/**
 	 * Apply operation to the image. The input may be modified (and the operation applied in-place), 
@@ -95,6 +101,22 @@ public interface ImageOp {
 	 */
 	public default PixelType getOutputType(PixelType inputType) {
 		return inputType;
+	}
+	
+	/**
+	 * Default implementation returns an empty list; classes should override this if necessary.
+	 */
+	@Override
+	public default Collection<URI> getUris() throws IOException {
+		return Collections.emptyList();
+	}
+	
+	/**
+	 * Default implementation makes no changes; classes should override this if necessary.
+	 */
+	@Override
+	public default boolean updateUris(Map<URI, URI>  replacements) throws IOException {
+		return false;
 	}
 
 }